requirements.txt
boto3==1.24.38
botocore==1.27.38
click==8.1.3
Flask==2.1.2
flask-talisman==1.0.0
Flask-WTF==1.0.1
gunicorn==20.1.0
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.11
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
